Test that a read of nonzero bytes from an InMemFile does not return EOF in the same read op.
